backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
psycopg2-binary==2.9.7
SQLAlchemy==1.4.49
alembic==1.11.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.9
python-multipart==0.0.6
python-dotenv==1.0.0
pytest==7.4.0
httpx==0.24.1
python-socketio==5.8.0
